+++ CHANGELOG.md
@@ -83,6 +83,7 @@
  - [x] Make notifications change rounding
  - [x] Fix hub notification not showing up
  - [x] Fix 3kh0 theme placeholder
+ - [x] Obfuscate notifications
 
 ### Roadmap
 The roadmap has moved [here](https://github.com/orgs/Metallic-Web/projects/1/views/1).